+++ src/main/kotlin/com/atlassian/performance/tools/awsinfrastructure/api/storage/ApplicationStorage.kt
@@ -3,6 +3,7 @@ package com.atlassian.performance.tools.awsinfrastructure.api.storage
 import com.atlassian.performance.tools.awsinfrastructure.AwsCli
 import com.atlassian.performance.tools.ssh.api.SshConnection
 
+@Deprecated("Use `com.atlassian.performance.tools.infrastructure.api.storage.ProductDistribution` instead.")
 interface ApplicationStorage {
     val possibleLocations: List<S3Artifact>