+++ src/main/kotlin/com/atlassian/performance/tools/awsinfrastructure/jira/StandaloneNodeFormula.kt
@@ -3,13 +3,13 @@ package com.atlassian.performance.tools.awsinfrastructure.jira
 import com.atlassian.performance.tools.aws.api.Storage
 import com.atlassian.performance.tools.awsinfrastructure.AwsCli
 import com.atlassian.performance.tools.awsinfrastructure.api.hardware.Computer
-import com.atlassian.performance.tools.awsinfrastructure.api.storage.ApplicationStorage
 import com.atlassian.performance.tools.infrastructure.api.Sed
 import com.atlassian.performance.tools.infrastructure.api.jira.JiraGcLog
 import com.atlassian.performance.tools.infrastructure.api.jira.JiraHomeSource
 import com.atlassian.performance.tools.infrastructure.api.jira.JiraNodeConfig
 import com.atlassian.performance.tools.infrastructure.api.jira.SetenvSh
 import com.atlassian.performance.tools.infrastructure.api.os.Ubuntu
+import com.atlassian.performance.tools.infrastructure.api.distribution.ProductDistribution
 import com.atlassian.performance.tools.jvmtasks.api.Backoff
 import com.atlassian.performance.tools.jvmtasks.api.IdempotentAction
 import com.atlassian.performance.tools.jvmtasks.api.TaskTimer
@@ -24,7 +24,7 @@ internal class StandaloneNodeFormula(
     private val pluginsTransport: Storage,
     private val resultsTransport: Storage,
     private val databaseIp: String,
-    private val application: ApplicationStorage,
+    private val productDistribution: ProductDistribution,
     private val ssh: Ssh,
     private val config: JiraNodeConfig,
     private val computer: Computer
@@ -40,14 +40,11 @@ internal class StandaloneNodeFormula(
 
         ssh.newConnection().use { connection ->
             computer.setUp(connection)
-            val jiraArchiveName = application.download(connection, ".")
+            val unpackedProduct = productDistribution.install(connection, ".")
             val jiraHome = TaskTimer.time("download Jira home") {
                 jiraHomeSource.download(connection)
             }
-            val unpackedProduct = getUnpackedProductName(connection, jiraArchiveName)
-
             replaceDbconfigUrl(connection, "$jiraHome/dbconfig.xml")
-            connection.execute("tar -xzf $jiraArchiveName", Duration.ofMinutes(1))
             SetenvSh(unpackedProduct).setup(
                 connection = connection,
                 config = config,
@@ -56,7 +53,10 @@ internal class StandaloneNodeFormula(
             )
             connection.execute("echo jira.home=`realpath $jiraHome` > $unpackedProduct/atlassian-jira/WEB-INF/classes/jira-application.properties")
             connection.execute("echo jira.autoexport=false > $jiraHome/jira-config.properties")
-            downloadMysqlConnector("https://dev.mysql.com/get/Downloads/Connector-J/mysql-connector-java-5.1.40.tar.gz", connection)
+            downloadMysqlConnector(
+                "https://dev.mysql.com/get/Downloads/Connector-J/mysql-connector-java-5.1.40.tar.gz",
+                connection
+            )
             connection.execute("tar -xzf mysql-connector-java-5.1.40.tar.gz")
             connection.execute("cp mysql-connector-java-5.1.40/mysql-connector-java-5.1.40-bin.jar $unpackedProduct/lib")
             AwsCli().download(pluginsTransport.location, connection, target = "$jiraHome/plugins/installed-plugins")
@@ -85,17 +85,6 @@ internal class StandaloneNodeFormula(
                 profiler = config.profiler
             )
         }
-    }
-
-    private fun getUnpackedProductName(
-        connection: SshConnection,
-        archiveName: String
-    ): String {
-        return connection
-            .execute("tar -tf $archiveName | head -n 1", timeout = Duration.ofMinutes(1))
-            .output
-            .split("/")
-            .first()
     }
 
     private fun replaceDbconfigUrl(